Finish dictionary JS logic, restructured vocab
Done with dictionary filters & sorting, changed from vocab to dict
