+++ helper/dictionary/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("dict", __name__)
+
+from helper.dictionary import routes
