+++ helper/dictionary/initial_load/lessons.json
