+++ helper/dictionary/routes.py
@@ -1,14 +1,12 @@
-from flask import render_template, request, url_for, current_app, redirect
-from helper.vocabulary import bp
-from helper.vocabulary.forms import TypeSelectors
+from flask import render_template, request, url_for, current_app
+from helper.dictionary import bp
 from helper.models import Entry
 
 
 @bp.route("/dictionary")
 def dictionary():
-    form = TypeSelectors()
     args = request.args
-    sorting = args.get("sorting", type=str)
+    sorting = args.get("srt", type=str)
     page = args.get("page", 1, type=int)
 
     noun = args.get("noun", type=str)
@@ -17,32 +15,23 @@ def dictionary():
 
     type_set = []
 
-    if noun == "y":
+    if noun == "true":
         type_set.append("noun")
-    else:
-        noun = "n"
-        form.noun.data = False
-    if phrase == "y":
+    if phrase == "true":
         type_set.append("phrase")
-    else:
-        phrase = "n"
-        form.phrase.data = False
-    if misc == "y":
+    if misc == "true":
         type_set.append("misc")
-    else:
-        misc = "n"
-        form.misc.data = False
 
     entries = Entry.get_entries(e_type=type_set, lst=False, sorting=sorting).paginate(
         page, current_app.config["DICT_ITEMS_PER_PAGE"], False
     )
     next_url = (
         url_for(
-            "vocab.dictionary",
+            "dict.dictionary",
             noun=noun,
             phrase=phrase,
             misc=misc,
-            sorting=sorting,
+            srt=sorting,
             page=entries.next_num,
         )
         if entries.has_next
@@ -50,41 +39,40 @@ def dictionary():
     )
     prev_url = (
         url_for(
-            "vocab.dictionary",
+            "dict.dictionary",
             noun=noun,
             phrase=phrase,
             misc=misc,
-            sorting=sorting,
+            srt=sorting,
             page=entries.prev_num,
         )
         if entries.has_prev
         else None
     )
     first_url = url_for(
-        "vocab.dictionary", noun=noun, phrase=phrase, misc=misc, sorting=sorting, page=1
+        "dict.dictionary", noun=noun, phrase=phrase, misc=misc, srt=sorting, page=1
     )
     last_url = url_for(
-        "vocab.dictionary",
+        "dict.dictionary",
         noun=noun,
         phrase=phrase,
         misc=misc,
-        sorting=sorting,
+        srt=sorting,
         page=entries.pages,
     )
 
     form_entries = [e.format_entry() for e in entries.items]
 
     return render_template(
-        "vocabulary/dictionary.html",
+        "dict/dictionary.html",
         title="Dictionary",
         entries=form_entries,
         next_url=next_url,
         prev_url=prev_url,
         first_url=first_url,
         last_url=last_url,
-        form=form,
-        sorting=sorting,
         noun=noun,
         phrase=phrase,
         misc=misc,
+        srt=sorting,
     )
