+++ helper/templates/index.html
@@ -2,6 +2,6 @@
 {% block app_content %}
 <div class="page-header">
     <h1>German Language Learning Helper</h1>
-    <h2>Site to help with the DW course <em>Nico's Weg</em></h2>
+    <p class="lead">Site to help with the DW course <em>Nico's Weg</em></p>
 </div>
 {% endblock %}
