--- helper/templates/vocabulary/dictionary.html
@@ -1,53 +1,0 @@
-{% extends "base.html" %}
-{% block app_content %}
-<h1>Dictionary</h1>
-<div class="dict-filters">
-    <form action="" method="GET" novalidate>
-        <p>
-            <span> {{ form.noun() }} {{ form.noun.label }} </span>
-            <span> {{ form.phrase() }} {{ form.phrase.label }}</span>
-            <span> {{ form.misc() }} {{ form.misc.label }}</span>
-            <span> {{ form.sorting.label }} {{ form.sorting() }}</span>
-            <span> {{ form.submit() }}</span>
-        </p>
-    </form>
-</div>
-<hr>
-<div class="container">
-    <table class="table table-hover">
-        <thead>
-            <tr>
-                <th scope="col">Entry</th>
-                <th scope="col">Translation</th>
-                <th scope="col">Base (nouns only)</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for entry in entries %}
-            <tr>
-                <td>{{ entry["entry"] }}</td>
-                <td>{{ entry["translation"] }}</td>
-                <td>{{ entry["base"] }}</td>
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-    <hr>
-</div>
-<nav aria-label="Page navigation example">
-  <ul class="pagination">
-    <li class="page-item"><a class="page-link" href="{{ first_url }}">First</a></li>
-    {% if prev_url %}
-        <li class="page-item"><a class="page-link" href="{{ prev_url }}">&larr;</a></li>
-    {% else %}
-        <li class="page-item disabled"><a class="page-link" href="#">&larr;</a></li>
-    {% endif %}
-    {% if next_url %}
-        <li class="page-item"><a class="page-link" href="{{ next_url }}">&rarr;</a></li>
-    {% else %}
-        <li class="page-item disabled"><a class="page-link" href="#">&rarr;</a></li>
-    {% endif %}
-    <li class="page-item"><a class="page-link" href="{{ last_url }}">Last</a></li>
-  </ul>
-</nav>
-{% endblock %}
