--- helper/vocabulary/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint("vocab", __name__)
-
-from helper.vocabulary import routes
